move common functionality over to utils
